# Make convolution JIT function more efficient for prime fields