… tests/fields/data/GF(109987^4)/convolve.pkl
… tests/fields/data/GF(2)/convolve.pkl
… tests/fields/data/GF(2147483647)/convolve.pkl
… tests/fields/data/GF(2^100)/convolve.pkl
… tests/fields/data/GF(2^2)/convolve.pkl
… tests/fields/data/GF(2^3)/convolve.pkl
… tests/fields/data/GF(2^32)/convolve.pkl
… tests/fields/data/GF(2^8)/convolve.pkl
… tests/fields/data/GF(2^8, 283, 19)/convolve.pkl
… tests/fields/data/GF(31)/convolve.pkl
… tests/fields/data/GF(3191)/convolve.pkl
… tests/fields/data/GF(36893488147419103183)/convolve.pkl
… tests/fields/data/GF(5)/convolve.pkl
… tests/fields/data/GF(7)/convolve.pkl
… tests/fields/data/GF(7^3)/convolve.pkl
… tests/fields/data/GF(7^3, 643, 244)/convolve.pkl