Convolution in `GF(p)` can be efficiently performed using NumPy's native `np.convolve()` and then reduced modulo p. I previously did this when directly invoking the convolution (ie in the override of `np.convolve()`). However, that optimization was not added in the JIT-ed convolution function (ie the one called by other JIT routines). This PR adds the optimization inside the JIT-able implementation so all callers receive the benefit.

### Before

```python
In [1]: import numpy as np

In [2]: import galois

In [3]: GF = galois.GF(2424113)

In [4]: x = GF.Random(10_000, seed=1)

In [5]: y = GF.Random(10_000, seed=2)

In [6]: %timeit np.convolve(x, y)
45.7 ms ± 93.8 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [7]: %timeit GF._convolve.function(x.astype(np.int64), y.astype(np.int64))
269 ms ± 697 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### After

```python
In [1]: import numpy as np

In [2]: import galois

In [3]: GF = galois.GF(2424113)

In [4]: x = GF.Random(10_000, seed=1)

In [5]: y = GF.Random(10_000, seed=2)

In [6]: %timeit np.convolve(x, y)
50 ms ± 4.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [7]: %timeit GF._convolve.function(x.astype(np.int64), y.astype(np.int64))
47.8 ms ± 171 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```